Add logic to code verification prompt and improve update error handling
